test out new redis dockerfile for drone

+++ src/services/postgres/postgres.js
@@ -1,18 +1,37 @@
 /* @flow */
 
 import knex from 'knex';
+import * as objection from 'objection';
+import * as objectionSoftDelete from 'objection-softdelete';
 import config from '../../config';
 
-const db = knex({
+let db;
+
+const knexOpts = {
   client: 'pg',
   connection: config.db.url,
   migrations: {
     tableName: 'migrations',
   },
   debug: config.db.debug,
-});
+};
+
+function connect() {
+  if (!db) {
+    db = knex(knexOpts);
+    const { Model } = objection;
+    // $FlowIssue
+    Model.knex(db);
+    objectionSoftDelete.register(objection);
+  }
+
+  return db;
+}
 
 async function disconnect(db: Object) {
+  if (!db) {
+    return;
+  }
   try {
     await db.destroy();
   } catch (err) {
@@ -20,6 +39,4 @@ async function disconnect(db: Object) {
   }
 }
 
-export default db;
-
-export { db, disconnect };
+export { connect, disconnect };
